Separate masthead background from masthead content
